+++ src/app/(light)/dashboard/org/[username]/asset/[id]/page.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next';
 
-import PDAPage, { getPDA } from '@/app/(light)/dashboard/user/asset/[id]/page';
+import PDAPage from '@/app/(light)/dashboard/user/asset/[id]/page';
+import { getPDA } from '@/services/server-functions/pda';
 
 export async function generateMetadata({
   params,
