+++ src/app/(light)/dashboard/user/asset/[id]/page.tsx
@@ -3,21 +3,10 @@ import { Metadata } from 'next';
 import BackButton from '@/components/buttons/back-button/back-button';
 import TopBarContainer from '@/components/containers/top-bar-container/top-bar-container';
 import routes from '@/constants/routes';
-import { getPrivateApi } from '@/services/protocol/api';
-import { PdaQuery } from '@/services/protocol/types';
+import { getPDA } from '@/services/server-functions/pda';
 import { getCurrentOrg } from '@/utils/currentOrg';
 
 import PDAItem from './components/pda-item';
-
-export const getPDA = async (id: string): Promise<PdaQuery['PDA'] | null> => {
-  const privateApi = await getPrivateApi();
-  if (!privateApi) {
-    return null;
-  }
-
-  const pda = (await privateApi.pda({ id }))?.PDA;
-  return pda;
-};
 
 export async function generateMetadata({
   params,
