+++ src/services/server-functions/pda.tsx
@@ -1,0 +1,12 @@
+import { getPrivateApi } from '../protocol/api';
+import { PdaQuery } from '../protocol/types';
+
+export const getPDA = async (id: string): Promise<PdaQuery['PDA'] | null> => {
+  const privateApi = await getPrivateApi();
+  if (!privateApi) {
+    return null;
+  }
+
+  const pda = (await privateApi.pda({ id }))?.PDA;
+  return pda;
+};
